+++ api_merchant/api_merchant/src/main/java/com/merchant/api/model/payload/request/OrderItemsReq.java
@@ -9,14 +9,14 @@ import lombok.ToString;
 @ToString
 public class OrderItemsReq {
 
-	private String itemId;
-	private String itemImageUrl;
-	private String itemName;
-	private Double itemPrice;
-	private Integer itemQuantity;
-	private String itemType;
-	private String itemUrl;
-	private String sellerBadge;
-	private String sellerName;
-	private String sellerId;
+    private String itemId;
+    private String itemImageUrl;
+    private String itemName;
+    private Integer itemPrice;
+    private Integer itemQuantity;
+    private String itemType;
+    private String itemUrl;
+    private String sellerBadge;
+    private String sellerName;
+    private String sellerId;
 }
